backend/context_ai/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "context_ai"
version = "0.1.0"
description = "Context AI RAG + Agent Service"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
	"fastapi==0.115.0",
	"uvicorn[standard]==0.30.6",
	"pymongo==4.8.0",
	"python-dotenv==1.0.1",
	"pydantic==2.9.2",
	"pydantic-settings==2.5.2",
	"sentence-transformers==3.1.1",
	"faiss-cpu>=1.9.0",
	"PyMuPDF==1.24.9",
	"rank-bm25==0.2.2",
	"numpy==2.1.3",
	"scikit-learn==1.5.2",
	"tiktoken==0.7.0",
	"transformers==4.45.2",
	"accelerate==0.34.2",
	"httpx==0.27.2",
	"pytest==8.3.3",
]

[tool.setuptools.packages.find]
where = ["src"]
